requirements.txt
onnx>=1.14.0,<1.16.0
numpy<2
torch>=2.0.0
coremltools>=8.3
onnx-simplifier>=0.4.36
